About
We have formulated an econometric model where we estimated the parameters by the following methods:
Pooled OLS, fixed effects, first differences and random effects. For these methods to be valid the conditions
have to be met. Through various tests it can be concluded that the conditions for pooled OLS, fixed
effects and random effects doesn’t hold. Vice versa, we learned that the conditions for first differences to
be valid were met. Although, it has to be stated that there is no strict exogeneity but we have argued that
FD consists of more exogeneity than FE. Thus, it can be concluded that, with the given data, the firms
don’t have CRS but diminishing returns since we saw in figure 3 that ˆ βFD K = 0.0630 and ˆ βFD L = 0.5487.